zephyr: hooks: fix compiler warning
Fixes:

/mnt/host/source/src/platform/ec/zephyr/shim/src/hooks.c:43:5: warning:
format '%x' expects argument of type 'unsigned int', but argument 3 has
type 'void (*)(void)' [-Wformat=]

BUG=none
BRANCH=none
TEST=build and run on volteer

+++ zephyr/shim/src/hooks.c
@@ -42,7 +42,7 @@ int hook_call_deferred(const struct deferred_data *data, int us)
 			cprintf(CC_HOOK,
 				"Warning: deferred call not submitted, "
 				"routine=0x%08x, err=%d",
-				non_const->routine, rv);
+				(unsigned int)non_const->routine, rv);
 	} else {
 		return EC_ERROR_PARAM2;
 	}
